[feat] 매일지출내역 저장 기능 구현

+++ src/main/java/com/zerobase/accountbook/domain/dailypayments/DailyPayments.java
@@ -24,7 +24,7 @@ public class DailyPayments {
     @JoinColumn(name = "MEMBER_ID")
     private Member member;
 
-    private Long paidAmount;
+    private Integer paidAmount;
 
     private String paidWhere;
